+++ SharpMap/Styles/Offset.cs
@@ -7,7 +7,7 @@ namespace SharpMap.Styles
     //I prefer to use an offsetX and offsetY. PDD.
     public class Offset
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
