+++ SharpMap/Styles/Pen.cs
@@ -5,7 +5,7 @@ namespace SharpMap.Styles
 {
     public class Pen
     {
-        public float Width { get; set; }
+        public double Width { get; set; }
         public Color Color { get; set; }
     }
 }
